backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
Pillow==10.4.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
zipp==3.23.0
